reorganized the registration package and resolved import issues. now on to testing it!
